Refactor files in 'src/app/twinclass/[twinClassId]'
