+++ src/app/twinclass/[twinClassId]/views/index.ts
@@ -1,0 +1,4 @@
+export * from "./twin-class-fields";
+export * from "./twin-class-general";
+export * from "./twin-class-statuses";
+export * from "./twin-class-twinflow";
