+++ src/app/twinclass/[twinClassId]/views/twin-class-fields/index.ts
@@ -1,0 +1,1 @@
+export * from "./twin-class-fields";
