+++ src/app/twinclass/[twinClassId]/views/twin-class-statuses/index.ts
@@ -1,0 +1,1 @@
+export * from "./twin-class-statuses";
