+++ src/app/twinclass/[twinClassId]/views/twin-class-twinflow/index.ts
@@ -1,0 +1,1 @@
+export * from "./twin-class-twinflows";
